Update README.md
Project was about predicting the percentage of marks that a student is expected to score based on the number of hours they studied. This is a simple linear regression task. My role as a data science intern was to write a function to make the prediction of scores using the linear regression model. Technologies used in the above projects: Python, GitHub, Jupiter Notebook, Supervised ML (Linear Regression), Pandas, NumPy.